pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-term-exec"
version = "0.1.0"
description = "Trigger a local AI via hotkey to generate and run shell commands from a TUI prompt."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name = "AI Assistant", email = "noreply@example.com" }
]
dependencies = [
  "typer>=0.12.3",
  "prompt_toolkit>=3.0.47",
  "requests>=2.31.0",
  "pynput>=1.7.6",
  "PyYAML>=6.0.1",
]

[project.scripts]
ai-term-exec = "ai_term_exec.cli:app"

[tool.setuptools]
packages = ["ai_term_exec"]

